refactor(type): 인풋 타입 사용 필드로 위치 수정
